Attempt to fix STR #2550 to make fl_text_extents work in Xlib-only (i.e. non-XFT) X11 builds.
This appears to be working now, though I can not test it fully, so we
need more feedback.
In particular the setting of the dx param is certainly wrong, but 
works fine in general cases. I think.
Also, I suspect it may behave badly in the face of RtoL text rendering
but I think we have other problems in that area anyway...