Porting back changes of branch-1.3-porting.
The old behavior is retained, but the files have been modified
to match the new versions in the porting branch as much as possible.

The porting branch has its own (modified) version. Windows executables
built with the new IDE's (generated by CMake) don't have the trailing 'd'
in Debug mode.