Fix Windows IDE solutions VisualC2008 and VisualC2010.
  Add proper quoting to build properly if the FLTK source directory
  is located in a path that contains spaces.

Note: this is a source tar file update that is only necessary if you
build one of the above mentioned Visual Studio IDE solutions in a path
with spaces. Nothing else changed, no documentation updates.

See also important notes below.

------------------------------------------------------------------------

Note to svn (subversion) users: This branch (branch-1.3) is finished
since the release of FLTK 1.3.4.

Only very serious bug fixes will be available in FLTK 1.3 if FLTK 1.4
is not yet released or not (yet) stable.

The new development branch is now branch-1.4 -- please checkout a new,
clean svn working copy if you want to get branch-1.4.

------------------------------------------------------------------------

Note to Git users: FLTK's svn branch-1.3 has been the 'master' branch
in our Git repository, but this branch is finished since the release
of FLTK 1.3.4.

Only very serious bug fixes will be available in the new Git 'branch-1.3'
(this branch) if FLTK 1.4 is not yet released or not (yet) stable.

The new development branch is now branch-1.4 -- please checkout
branch-1.4 if you want to get the current FLTK 1.4 development branch.

The Git 'master' branch will stay as-is for some time, but in the
near future branch-1.4 will likely become the new 'master' branch.